lsm_tree.c  592  Info 734: Loss of precision (assignment) (31 bits to 7 bits)

+++ src/lsm/lsm_tree.c
@@ -589,7 +589,7 @@ __lsm_tree_open(WT_SESSION_IMPL *session,
 	 * with getting handles exclusive.
 	 */
 	lsm_tree->refcnt = 1;
-	lsm_tree->exclusive = exclusive;
+	lsm_tree->exclusive = (int8_t)exclusive;
 	lsm_tree->queue_ref = 0;
 
 	/* Set a flush timestamp as a baseline. */
